OpenBlastGenerator
