+++ Source/Frontend/UI/Components/Engine Config/CorruptionEngineForm.cs
@@ -361,17 +361,6 @@ namespace RTCV.UI
             }
         }
 
-        private void OpenBlastGenerator(object sender, EventArgs e)
-        {
-            if (S.GET<BlastGeneratorForm>() != null)
-            {
-                S.GET<BlastGeneratorForm>().Close();
-            }
-
-            S.SET(new BlastGeneratorForm());
-            S.GET<BlastGeneratorForm>().LoadNoStashKey();
-        }
-
         private void UpdateBlastType(object sender, EventArgs e)
         {
             switch (gbNightmareEngine.cbBlastType.SelectedItem.ToString())
