+++ Source/Frontend/UI/Components/Engine Config/Engines/BlastGeneratorEngine.cs
@@ -1,10 +1,24 @@
 namespace RTCV.UI.Components.EngineConfig.Engines
 {
-    public partial class BlastGeneratorEngine : EngineConfigControl
+    using System;
+    using RTCV.Common;
+
+    internal partial class BlastGeneratorEngine : EngineConfigControl
     {
-        public BlastGeneratorEngine()
+        internal BlastGeneratorEngine()
         {
             InitializeComponent();
+        }
+
+        private void OpenBlastGenerator(object sender, EventArgs e)
+        {
+            if (S.GET<BlastGeneratorForm>() != null)
+            {
+                S.GET<BlastGeneratorForm>().Close();
+            }
+
+            S.SET(new BlastGeneratorForm());
+            S.GET<BlastGeneratorForm>().LoadNoStashKey();
         }
     }
 }
